+++ operators/510_list.c
@@ -17,7 +17,7 @@ mud_object_t * _mud_op_list_reverse_evaluate(mud_expr_evaluator_t * evaluator) {
   mud_object_t * ret = ME_ARG(0);
   mud_list_t * old_list = (mud_list_t *)ret->ptr;
   mud_list_t * new_list = mud_list_alloc();
-  for ( unsigned i = old_list->count; i > 0; ) {
+  for ( size_t i = old_list->count; i > 0; ) {
     mud_list_append(new_list, old_list->objects[--i]);
   }
   mud_list_free(old_list);
